Update support-for-windows-10.md
I am proposing three specific sets of changes to this topic. The first proposed change is to stop using the term "version" in reference to Windows 10, replacing it with the term "release" (with release used as a noun and not a verb).  

The purpose of this change is to insure we do not imply that Windows 10 has "multiple versions" when using the word “version” with Windows 10, but instead insure we are enforcing the idea that Windows 10 is in fact the last "version" of Windows (there is no "Windows 11" coming), however new capabilities and features are added to Windows over time through Feature Update "releases", where a release "becomes available" or "is generally available" (“available” is used to avoid writing that a “release” (noun) is being "released" (verb)).

The second proposed change is to update the chart and references under the “Windows as a Configuration Manager client” section to reference the support dates for the SCCM Current Branch (CB) version releases, and to update the icons for SCCM “1606” now that it has reached it’s “End of Support milestone”.  

These changes are realized through proposed changes to the icons for SCCM “1606” in the chart itself, as well as references to the actual end of Support milestone as text under each release described in the chart.  Additionally, a bullet referring readers to the Document Library topic titled “Support for System Center Configuration Manager current branch versions” is also added.  Since “support” of Windows 10 as a SCCM “client” depends on the status of each SCCM CB version release, I believe convenient reference from this topic to the “end of support” date topic is important.

The third proposed change is to remove the references to "CB" and "CBB" in relation to Windows 10 "releases". 

Documentation and discussions with Commercial Customers on the topic of Windows as a Service are moving away from the use of the "Current Branch (CB)" and "Current Branch for Business" (CBB) terminology.  Instead we (Microsoft Commercial Technical Field) are tasked with working to focus our customers on understanding that "CBB" is simply a development milestone in the life of a Windows 10 release, and they should be deploying each Semi-Annual Channel Feature Update release for Windows 10 using a process and timing that makes sense for their organization, rather than automatically having the organizational process being tied to waiting for the “CBB” development milestone just because it is titled “Current Branch for Business”. 

It is important we de-emphasize CBB as a deployment milestone for commercial customers, noting that CBB is actually a milestone in the code development of an individual Windows release that is based on a certain Quality Update being available and deployed.  We need customers to understand there is really no significant difference to Windows 10 code except for the installation of that Quality Update release that designates the CBB milestone (for example, Win 10 "1607" reaches "CBB" when the Quality Update detailed by KB 32000970 is deployed ). 

However, since SCCM CB release support for Windows 10 can be different before or after the Windows 10 release code reaches the CBB development milestone, I am proposing that any references made to CBB are avoided, and instead a reference made to the Quality Update for the Windows 10 release that brings the code to “CBB”.  This could appear in future versions of this document as a note, such as "NOTE: Backwards Compatibility Support for Windows 10 “1607” ends once Windows 10 "1607" is serviced using the Quality Update released under KB 32000970, made generally available on 2016-11-08)

I hope these proposed changes will be considered.