Fixed inaccuracy (Cisco IPsec for iOS, not macOS)
I replaced "macOS" with "iOS" for the Cisco (IPsec) VPN profile type feature as this functionality was actually added for iOS only; the doc it points to is correct. I created this new pull request in the private branch (and will close my PR in the main branch) at @aczechowski's suggestion.