Undocumented AdvertFlags
It seems we have some undocumented AdvertFlags. It seems these are for some deployment options added to the product later, e.g.:

19 - "Allow task sequence to run for client on the Internet"
28 - Deployment Settings\Make available for the following - "Only media and PXE"
29 - Deployment Settings\Make available for the following - "Only media and PXE (hidden)"

The above flags and corresponding deployment options are based on my observations in the lab and some DL discussions, but should be double-checked and also validate if there are any addtiional un-documented flags.