Create importmachineentrytomultiplecollections-method-in-class-sms_site.md
This document was missing from the library. It's just a copy of the ImportMachineEntry Method importmachineentry-method-in-class-sms_site.md document, with the method name, CollectionID string changed to CollectionIDs string[].  Not sure if I'm supposed to leave the header info in there or not. This was also reported in issue 332 https://github.com/MicrosoftDocs/SCCMdocs/issues/332.