Update plan-for-and-configure-application-management.md
Some customers are confused as there is no "none" in the client settings.  It only changes to "none" when you select "Automatically  Detect" in the dropdown menu, so it appears at first sight that there is an issue due to not having an option of "None".  Also pitched in some CM to MEM changes but not sure if it's too early for that.